Minor change to reading cmd
Now the readCmdFromBluetooth method will read from the serial port and save each byte to string but would not return it if "enter" is not read. The next time this method is run, it will add read bytes to the string. So, the command will only be executed if that method read "enter".