Refactored Node Id code
I took the node id code and refactored it so that we have checkNodes method to  increment nodes counts, checkToTurn method to set motor speeds according to the direction and node to be turned on and waitWhileTurning method which loops while the robot is not centered on the new line.
I also changed the centered method to remove the boolean array and directly checked each sensor for finding line in if statement.
Also, there are minor variable name changes to make the code more readable.